requirements.txt
pandas==1.5.3
Pillow==10.0.0
streamlit>=1.20.0
